requirements.txt
streamlit
pandas
requests
python-docx
gspread
oauth2client
